+++ CHANGES.md
@@ -6,6 +6,7 @@
   service and not providing HTTP-specific options
 * Add `IncludedTraffic`, `OutgoingTraffic` and `IngoingTraffic` fields to `LoadBalancer`
 * Add `ChangeType()` method to the Load Balancer client
+* Fix retrying of requests that contain a body
 
 ## v1.18.2